[ Infrastructure ] Change DateTimeTicks in Log model from uint to ulong

+++ Biovation.Brands.Suprema/BioStarServer.cs
@@ -745,7 +745,7 @@ namespace Biovation.Brands.Suprema
                     receivedLog.DeviceCode = device.Code;
                     receivedLog.EventLog = _supremaCodeMappings.GetLogEventGenericLookup(logRecord.Event) ??
                                            new Lookup { Code = logRecord.Event.ToString() };
-                    receivedLog.DateTimeTicks = (uint)logRecord.eventTime;
+                    receivedLog.DateTimeTicks = (ulong)logRecord.eventTime;
                     receivedLog.InOutMode = device.DeviceTypeId;
                     receivedLog.Reserved = logRecord.reserved1;
                     receivedLog.TnaEvent = logRecord.tnaKey;
@@ -981,7 +981,7 @@ namespace Biovation.Brands.Suprema
                     DeviceId = device.DeviceId,
                     DeviceCode = device.Code,
                     EventLog = _supremaCodeMappings.GetLogEventGenericLookup(logRecord.Event) ?? new Lookup { Code = logRecord.Event.ToString() },
-                    DateTimeTicks = (uint)logRecord.eventTime,
+                    DateTimeTicks = (ulong)logRecord.eventTime,
                     InOutMode = device.DeviceTypeId,
                     Reserved = logRecord.reserved,
                     TnaEvent = logRecord.tnaEvent,
@@ -1187,7 +1187,7 @@ namespace Biovation.Brands.Suprema
 
                     var faceLog = new Log
                     {
-                        DateTimeTicks = (uint)imageLogHdr.eventTime,
+                        DateTimeTicks = (ulong)imageLogHdr.eventTime,
                         //EventLog =SupremaLogService Convert.ToInt32(imageLogHdr.Event),
                         EventLog = _supremaCodeMappings.GetLogEventGenericLookup(imageLogHdr.Event),
                         DeviceId = (int)deviceId,
